+++ ios/Classes/Modules/FInjector.m
@@ -12,9 +12,23 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface FInjector () <SNRInjectorInAppMessageDelegate>
 
+@property (nonatomic, strong) NSMutableDictionary<NSString *, SNRInAppCustomMethodCompletion *> *pendingInAppCustomMethodCompletions;
+
 @end
 
 @implementation FInjector
+
+#pragma mark - Lifecycle
+
+- (instancetype)init {
+    self = [super init];
+
+    if (self) {
+        _pendingInAppCustomMethodCompletions = [NSMutableDictionary dictionary];
+    }
+
+    return self;
+}
 
 #pragma mark - Public
 
@@ -23,13 +37,35 @@ NS_ASSUME_NONNULL_BEGIN
 }
 
 - (void)handleMethodCall:(FlutterMethodCall*)call result:(FlutterResult)result calledMethod:(NSString *)calledMethod {
-    if ([calledMethod isEqualToString:@"closeInAppMessage"]) {
+    if ([calledMethod isEqualToString:@"setInAppContext"]) {
+        [self setInAppContext:call result:result];
+    } else if ([calledMethod isEqualToString:@"notifyInAppContextChange"]) {
+        [self notifyInAppContextChange:call result:result];
+    } else if ([calledMethod isEqualToString:@"closeInAppMessage"]) {
         [self closeInAppMessage:call result:result];
+    } else if ([calledMethod isEqualToString:@"handleOpenUrlBySDK"]) {
+        [self handleOpenUrlBySDK:call result:result];
     } else if ([calledMethod isEqualToString:@"handleDeepLinkBySDK"]) {
         [self handleDeepLinkBySDK:call result:result];
-    } else if ([calledMethod isEqualToString:@"handleDeepLinkBySDK"]) {
-        [self handleDeepLinkBySDK:call result:result];
+    }  else if ([calledMethod isEqualToString:@"resolveInAppCustomMethod"]) {
+        [self resolveInAppCustomMethod:call result:result];
     }
+}
+
+#pragma mark - Methods
+
+- (void)setInAppContext:(FlutterMethodCall *)call result:(FlutterResult)result {
+    id context = call.arguments;
+    if ([context isKindOfClass:[NSDictionary class]] == YES) {
+        [SNRInjector setInAppContext:context];
+    }
+
+    result([NSNumber numberWithBool:YES]);
+}
+
+- (void)notifyInAppContextChange:(FlutterMethodCall *)call result:(FlutterResult)result {
+    [SNRInjector notifyInAppContextChange];
+    result([NSNumber numberWithBool:YES]);
 }
 
 - (void)executeURLAction:(NSURL *)URL source:(SNRSyneriseSource)source {
@@ -46,8 +82,6 @@ NS_ASSUME_NONNULL_BEGIN
     }];
 }
 
-#pragma mark - Methods
-
 - (void)closeInAppMessage:(FlutterMethodCall *)call result:(FlutterResult)result {
     NSString *campaignHash = (NSString *)call.arguments;
     [SNRInjector closeInAppMessageWithCampaignHash:campaignHash];
@@ -56,7 +90,7 @@ NS_ASSUME_NONNULL_BEGIN
 - (void)handleOpenUrlBySDK:(FlutterMethodCall *)call result:(FlutterResult)result {
     NSString *urlString = (NSString *)call.arguments;
     NSURL *URL = [NSURL URLWithString:urlString];
-    
+
     dispatch_async(dispatch_get_main_queue(), ^{
         if ([[UIApplication sharedApplication] canOpenURL:URL]) {
             if (@available(iOS 10, *)) {
@@ -71,16 +105,51 @@ NS_ASSUME_NONNULL_BEGIN
 - (void)handleDeepLinkBySDK:(FlutterMethodCall *)call result:(FlutterResult)result {
     NSString *deepLink = (NSString *)call.arguments;
     NSURL *deepLinkURL = [NSURL URLWithString:deepLink];
-    
+
     dispatch_async(dispatch_get_main_queue(), ^{
         if (@available(iOS 10, *)) {
             [[UIApplication sharedApplication] openURL:deepLinkURL options:@{} completionHandler:^(BOOL success) {
-                
+
             }];
         } else {
             [[UIApplication sharedApplication] openURL:deepLinkURL];
         }
     });
+}
+
+- (void)resolveInAppCustomMethod:(FlutterMethodCall *)call result:(FlutterResult)result {
+    NSDictionary *dictionary = call.arguments;
+    NSString *callId = [dictionary getStringForKey:@"callId"];
+    if (callId == nil) {
+        return;
+    }
+
+    BOOL isSuccess = [dictionary getBoolForKey:@"success"];
+    id methodResult = dictionary[@"result"];
+    if ([methodResult isKindOfClass:[NSNull class]] == YES) {
+        methodResult = nil;
+    }
+    NSString *error = [dictionary getStringForKey:@"error"];
+
+    SNRInAppCustomMethodCompletion *completion;
+    @synchronized (self) {
+        completion = self.pendingInAppCustomMethodCompletions[callId];
+        [self.pendingInAppCustomMethodCompletions removeObjectForKey:callId];
+    }
+
+    if (completion == nil) {
+        return;
+    }
+
+    if (isSuccess == YES) {
+        [completion success:methodResult];
+    } else {
+        if (error == nil) {
+          error = @"Unknown error.";
+        }
+
+        [completion failure:error];
+    }
 }
 
 #pragma mark - Dart Mapping
@@ -98,7 +167,7 @@ NS_ASSUME_NONNULL_BEGIN
 - (nullable NSDictionary *)dictionaryWithInAppMessageData:(nullable SNRInAppMessageData *)model {
     if (model != nil) {
         NSMutableDictionary *dictionary = [@{} mutableCopy];
-        
+
         [dictionary setString:model.campaignHash forKey:@"campaignHash"];
         [dictionary setString:model.variantIdentifier forKey:@"variantIdentifier"];
         [dictionary setDictionary:(model.additionalParameters ?: @{}) forKey:@"additionalParameters"];
@@ -106,7 +175,7 @@ NS_ASSUME_NONNULL_BEGIN
 
         return dictionary;
     }
-    
+
     return nil;
 }
 
@@ -124,6 +193,10 @@ NS_ASSUME_NONNULL_BEGIN
        @"data": [self dictionaryWithInAppMessageData:data]
     };
     [[FSyneriseManager sharedInstance].reverseChannel invokeMethod:@"Injector#InjectorInAppMessageListener#onHide" arguments:dictionary];
+}
+
+- (nullable NSDictionary *)SNR_inAppMessageContextIsNeeded:(SNRInAppMessageData *)data {
+    return nil;
 }
 
 - (void)SNR_inAppMessageHandledURLAction:(SNRInAppMessageData *)data url:(NSURL *)url {
@@ -149,6 +222,22 @@ NS_ASSUME_NONNULL_BEGIN
        @"parameters": parameters
     };
     [[FSyneriseManager sharedInstance].reverseChannel invokeMethod:@"Injector#InjectorInAppMessageListener#onCustomAction" arguments:dictionary];
+}
+
+- (void)SNR_inAppMessageHandledCustomMethod:(SNRInAppMessageData *)data name:(NSString *)name parameters:(NSDictionary *)parameters completion:(SNRInAppCustomMethodCompletion *)completion {
+    NSString *callId = [[NSUUID UUID] UUIDString];
+    @synchronized (self) {
+        self.pendingInAppCustomMethodCompletions[callId] = completion;
+    }
+
+    NSDictionary *dataDictionary = [self dictionaryWithInAppMessageData:data];
+    NSDictionary *dictionary = @{
+       @"callId": callId,
+       @"name": name,
+       @"parameters": parameters != nil ? parameters : @{},
+       @"data": dataDictionary != nil ? dataDictionary : @{}
+    };
+    [[FSyneriseManager sharedInstance].reverseChannel invokeMethod:@"Injector#InjectorInAppMessageListener#onCustomMethod" arguments:dictionary];
 }
 
 @end