+++ lib/main/index.dart
@@ -76,6 +76,7 @@ export 'package:synerise_flutter_sdk/enums/vouchers/voucher_code_status.dart';
 // INJECTOR
 export 'package:synerise_flutter_sdk/enums/injector/synerise_source.dart';
 export 'package:synerise_flutter_sdk/model/in_app/in_app_message_data.dart';
+export 'package:synerise_flutter_sdk/model/in_app/in_app_custom_method_completion.dart';
 
 // TRACKER
 export 'package:synerise_flutter_sdk/events/event.dart';